Use react-router-dom instead of @reach/router
I had a hard time figuring out a clean way to implement protected routes
and a login component that would redirect to the originally requested
url, like is easy with react-router. I like a lot about the way reach
router works, but for this purpose it was easier just to make the
switch.