Add Teams and Standings tabs to Divisions page
The idea with this change is that there are two things you might want to
do with a division. You're either trying to find your team, in which
case it's probably helpful to see them in some consistent, predictable
order (in this case, alphabetical), rather than in rank order, which
changes throughout the season.

The other thing you might want to do is to see the detailed standings of
the division, without having to navigate through a random team in that
division to see them.

Although this makes the page a little more cluttered, it might help with
those use cases a bit, and now we have the new tabbed interface on both
the Division and Team pages. We'll see what people think.