Add remove favorite button to each favorite entry
As part of this change I extracted NavListItem out of NavListLink so
that it can be used in more contexts, and made NavListLink just the
styled(Link) itself, where NavListItem is just a styled(li). There are
probably other places in the app where I can make use of NavListItem to
reduce some duplication, but I'll leave that for another time.