Updated to support the Team Sideline version of the website
The main changes were to replace match points and anything related to
them with average point differential, which is the second tier
tie-breaker on the new site.

The new site also doesn't list who will be reffing a match, and it
provides the location in a very abbreviated form (PA for Portland
Adventist Elementary), but it also includes a map link, so I've removed
the ref section and made the location name a link to the map so people
can at least see where the gym is if not devise its name from the
abbreviation.