Add context menu and sub finder
I think I eventually want this to be a navigable state so that if you
click on a team in the modal and then go back you will arrive back at
the modal. This will require a bit of refactoring, but that can be a
next step. I think this is probably good enough for a MVP.